+++ CHANGELOG.md
@@ -6,6 +6,33 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+## 0.0.13 (22th Sep, 2023)
+
+* Update project setting
+
+## 0.0.12 (22th Sep, 2023)
+
+### Fixed
+
+* No exception when log to file anymore
+
+## 0.0.11 (22th Sep, 2023)
+
+### Added
+
+* Allow to retry on failure
+* Save failed data to file
+
+### Fixed
+
+* Handle some SqlAlchemy async errors
+  * got Future <Future pending> attached to a different loop
+  * Command Out of Sync
+
+## 0.0.10 (21th Sep, 2023)
+
+* Revert 0.0.9
+
 ## 0.0.8 (7th Sep, 2023)
 
 ### Added